+++ CHANGELOG.md
@@ -10,7 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Breaking Changes
   * None
 * Added
-  * None
+  * `Authlogic::Session::Base.session_fixation_defense` - Reset the Rack
+    session ID after authentication, to protect against Session Fixation
+    attacks. (https://guides.rubyonrails.org/security.html#session-fixation)
+    Default: true
 * Fixed
   * None